# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Bootentry:
    properties:
      computers:
        items:
          $ref: '#/definitions/models.Computer'
        type: array
      created_at:
        type: string
      description:
        type: string
      files:
        items:
          $ref: '#/definitions/models.BootentryFile'
        type: array
      id:
        type: string
      ipxe_script:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.BootentryFile:
    properties:
      name:
        type: string
      protected:
        type: boolean
    type: object
  models.Computer:
    properties:
      asset:
        type: string
      bootentry_uuid:
        type: string
      build_arch:
        type: string
      created_at:
        type: string
      hostname:
        type: string
      id:
        type: string
      last_seen:
        type: string
      manufacturer:
        type: string
      name:
        type: string
      platform:
        type: string
      product:
        type: string
      serial:
        type: string
      tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      updated_at:
        type: string
      version:
        type: string
    type: object
  models.Error:
    properties:
      error:
        type: string
    type: object
  models.Ipxeaccount:
    properties:
      created_at:
        type: string
      is_admin:
        type: boolean
      last_login:
        type: string
      password:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.Tag:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Manage PXE boot
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ipxeblue API
  version: "0.1"
paths:
  /Bootentries/{username}:
    delete:
      consumes:
      - application/json
      description: Delete Bootentry
      parameters:
      - description: Bootentry UUID
        in: path
        maxLength: 36
        minLength: 36
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Failed to parse UUID
          schema:
            $ref: '#/definitions/models.Error'
        "404":
          description: Bootentry UUID not found
          schema:
            $ref: '#/definitions/models.Error'
      summary: Delete Bootentry
  /bootentries:
    get:
      consumes:
      - application/json
      description: List of Bootentry filtered or not
      parameters:
      - description: Offset
        in: query
        name: _start
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Bootentry'
            type: array
      summary: List Bootentries
    post:
      consumes:
      - application/json
      description: Create a Bootentry
      parameters:
      - description: json format Bootentry
        in: body
        name: bootentry
        required: true
        schema:
          $ref: '#/definitions/models.Bootentry'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Bootentry'
        "400":
          description: Failed to create account in DB
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Unmarshall error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Create Bootentry
  /bootentries/{id}:
    get:
      consumes:
      - application/json
      description: Get a Bootentry by Id
      parameters:
      - description: Bootentry UUID
        in: path
        maxLength: 36
        minLength: 36
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Bootentry'
        "404":
          description: Computer with uuid %s not found
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get Bootentry
  /bootentries/{username}:
    put:
      consumes:
      - application/json
      description: Update a Bootentry
      parameters:
      - description: Bootentry UUID
        in: path
        maxLength: 36
        minLength: 36
        name: uuid
        required: true
        type: string
      - description: json format of Bootentry
        in: body
        name: bootentry
        required: true
        schema:
          $ref: '#/definitions/models.Bootentry'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Bootentry'
        "400":
          description: Query uuid and uuid miss match
          schema:
            $ref: '#/definitions/models.Error'
        "404":
          description: Bootentry UUID not found
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Unmarshall error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Update Bootentry
  /computers:
    get:
      consumes:
      - application/json
      description: List of computers filtered or not
      parameters:
      - description: Offset
        in: query
        name: _start
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Computer'
            type: array
      summary: List computers
  /computers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a computer
      parameters:
      - description: Computer UUID
        in: path
        maxLength: 36
        minLength: 36
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Failed to parse UUID
          schema:
            $ref: '#/definitions/models.Error'
        "404":
          description: Can not find ID
          schema:
            $ref: '#/definitions/models.Error'
      summary: Delete computer
    get:
      consumes:
      - application/json
      description: Get a computer by Id
      parameters:
      - description: Computer UUID
        in: path
        maxLength: 36
        minLength: 36
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Computer'
        "404":
          description: Computer with uuid %s not found
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get computer
    put:
      consumes:
      - application/json
      description: Update a computer
      parameters:
      - description: Computer UUID
        in: path
        maxLength: 36
        minLength: 36
        name: id
        required: true
        type: string
      - description: json format computer
        in: body
        name: computer
        required: true
        schema:
          $ref: '#/definitions/models.Computer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Computer'
        "400":
          description: Query ID and UUID miss match
          schema:
            $ref: '#/definitions/models.Error'
        "404":
          description: Can not find ID
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Unmarshall error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Update computer
  /ipxeaccount:
    get:
      consumes:
      - application/json
      description: List of accounts for ipxe
      parameters:
      - description: Offset
        in: query
        name: _start
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Ipxeaccount'
            type: array
      summary: List iPXE account
    post:
      consumes:
      - application/json
      description: Create a iPXE account
      parameters:
      - description: json format iPXE account
        in: body
        name: ipxeaccount
        required: true
        schema:
          $ref: '#/definitions/models.Ipxeaccount'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Ipxeaccount'
        "400":
          description: Failed to create account in DB
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Unmarshall error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Create iPXE account
  /ipxeaccount/{username}:
    delete:
      consumes:
      - application/json
      description: Delete a iPXE account
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Ipxeaccount'
        "404":
          description: iPXE account not found
          schema:
            $ref: '#/definitions/models.Error'
      summary: Delete iPXE account
    get:
      consumes:
      - application/json
      description: Get iPXE account by username
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Ipxeaccount'
        "404":
          description: iPXE account not found
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get iPXE account
    put:
      consumes:
      - application/json
      description: Update a iPXE account
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: json format iPXE account
        in: body
        name: ipxeaccount
        required: true
        schema:
          $ref: '#/definitions/models.Ipxeaccount'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Ipxeaccount'
        "400":
          description: Query username and username miss match
          schema:
            $ref: '#/definitions/models.Error'
        "404":
          description: iPXE account not found
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Unmarshall error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Update iPXE account
swagger: "2.0"